refactor: move test files around
