+++ internal/cli/branch/modify_test.go
@@ -11,7 +11,13 @@ import (
 
 func TestModifyCommand(t *testing.T) {
 	t.Parallel()
-	binaryPath := getStackitBinary(t)
+	binaryPath := testhelpers.GetSharedBinaryPath()
+	if binaryPath == "" {
+		if err := testhelpers.GetBinaryError(); err != nil {
+			t.Fatalf("failed to build stackit binary: %v", err)
+		}
+		t.Fatal("stackit binary not built")
+	}
 
 	t.Run("modify amends commit with --all and --message", func(t *testing.T) {
 		t.Parallel()
