+++ internal/cli/navigation/parent_test.go
@@ -7,15 +7,14 @@ import (
 
 	"github.com/stretchr/testify/require"
 
-	"stackit.dev/stackit/internal/testhelper"
 	"stackit.dev/stackit/testhelpers"
 )
 
 func TestParentCommand(t *testing.T) {
 	t.Parallel()
-	binaryPath := testhelper.GetSharedBinaryPath()
+	binaryPath := testhelpers.GetSharedBinaryPath()
 	if binaryPath == "" {
-		if err := testhelper.GetBinaryError(); err != nil {
+		if err := testhelpers.GetBinaryError(); err != nil {
 			t.Fatalf("failed to build stackit binary: %v", err)
 		}
 		t.Fatal("stackit binary not built")
