+++ internal/cli/stack/reorder_test.go
@@ -13,7 +13,13 @@ import (
 
 func TestReorderCommand(t *testing.T) {
 	t.Parallel()
-	binaryPath := getStackitBinary(t)
+	binaryPath := testhelpers.GetSharedBinaryPath()
+	if binaryPath == "" {
+		if err := testhelpers.GetBinaryError(); err != nil {
+			t.Fatalf("failed to build stackit binary: %v", err)
+		}
+		t.Fatal("stackit binary not built")
+	}
 
 	t.Run("simple reorder: A -> B to B -> A", func(t *testing.T) {
 		t.Parallel()
