+++ internal/cli/stack/submit_test.go
@@ -13,7 +13,13 @@ import (
 func TestSubmitCommand(t *testing.T) {
 	t.Parallel()
 	// Build the stackit binary first
-	binaryPath := getStackitBinary(t)
+	binaryPath := testhelpers.GetSharedBinaryPath()
+	if binaryPath == "" {
+		if err := testhelpers.GetBinaryError(); err != nil {
+			t.Fatalf("failed to build stackit binary: %v", err)
+		}
+		t.Fatal("stackit binary not built")
+	}
 
 	t.Run("submit includes current branch in list", func(t *testing.T) {
 		t.Parallel()
