+++ internal/integration/main_test.go
@@ -4,15 +4,15 @@ package integration
 import (
 	"testing"
 
-	"stackit.dev/stackit/internal/testhelper"
+	"stackit.dev/stackit/testhelpers"
 )
 
 // getStackitBinary returns the path to the pre-built stackit binary.
 func getStackitBinary(t *testing.T) string {
 	t.Helper()
-	binaryPath := testhelper.GetSharedBinaryPath()
+	binaryPath := testhelpers.GetSharedBinaryPath()
 	if binaryPath == "" {
-		if err := testhelper.GetBinaryError(); err != nil {
+		if err := testhelpers.GetBinaryError(); err != nil {
 			t.Fatalf("failed to build stackit binary: %v", err)
 		}
 		t.Fatal("stackit binary not built")
